Replace Date to long for ticks.
- add more tests for tick conversion.
- add endpoints to #ticks processed.
- makes tickservices async. (need more test).

+++ src/main/java/io/cax/forex/ingester/repositories/TickRepository.java
@@ -8,6 +8,6 @@ import org.springframework.stereotype.Repository;
  * Created by cq on 15/4/16.
  */
 @Repository
-public interface TickRepository extends CrudRepository<Tick,Integer> {
+public interface TickRepository extends CrudRepository<Tick,Long> {
 
 }
